High bandwidth changes
Changed Send Message application to accept U32 to be uniform with Capture. Also modified NDL to accept message_count for send messages apps.

+++ sim/elvis/src/applications/send_message.rs
@@ -26,7 +26,7 @@ pub struct SendMessage {
     /// The machine that will receive the message
     remote_mac: Option<Mac>,
     /// The number of copies of the message to send
-    count: u16,
+    count: u32,
     /// The protocol to use in delivering the message
     transport: Transport,
 }
@@ -56,7 +56,7 @@ impl SendMessage {
     }
 
     /// The number of copies of the message to send
-    pub fn count(mut self, count: u16) -> Self {
+    pub fn count(mut self, count: u32) -> Self {
         self.count = count;
         self
     }
